+++ src/Domain/Cotor.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace IServ\ComposerToolsInstaller\Domain;
+
+final class Cotor
+{
+    /** @var string */
+    public const COMPOSER_EXTRA = 'cotor';
+
+    /** @var string */
+    public const COMPOSER_EXTRA_EXTENSIONS = 'extensions';
+
+    private function __construct()
+    {
+    }
+}