+++ src/Domain/Package.php
@@ -6,6 +6,8 @@ namespace IServ\ComposerToolsInstaller\Domain;
 
 final class Package
 {
+    public const DEFAULT_VERSION = '*';
+
     /** @var string */
     private $vendor;
 
@@ -15,15 +17,19 @@ final class Package
     /** @var string */
     private $version;
 
-    public function __construct(string $vendor, string $name, string $version = '*')
+    public function __construct(string $vendor, string $name, string $version = self::DEFAULT_VERSION)
     {
         $this->vendor = $vendor;
         $this->name = $name;
         $this->version = $version;
     }
 
-    public static function createFromComposerName(string $vendorName, string $version): self
+    public static function createFromComposerName(string $vendorName, string $version = self::DEFAULT_VERSION): self
     {
+        if (substr_count($vendorName, '/') !== 1) {
+            throw new \InvalidArgumentException(sprintf('%s is not valid composer package name!', $vendorName));
+        }
+
         [$vendor, $name] = explode('/', $vendorName, 2);
 
         return new self($vendor, $name, $version);